% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{age_pension}
\alias{age_pension}
\title{Vectorized age pension}
\usage{
age_pension(Assets, Income, HomeOwner, Age, Year)
}
\arguments{
\item{Assets}{Vector of assets.}

\item{Income}{Vector of income.}

\item{HomeOwner}{Vector of home-owner.}

\item{Age}{Vector of ages.}

\item{Year}{Vector of years}
}
\description{
This function is the same \code{\link{AgePension}} but vectorized. Perversely it should be faster than linear.
}
\details{
\code{age_pension} is a vectorized version of \code{\link{AgePension}}.
}
