% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AgePension}
\alias{AgePension}
\title{Age pension of singles.}
\arguments{
\item{assets}{Assessable assets.}

\item{income}{Assessable income.}

\item{age}{Age (integer).}

\item{year}{Year of entitlement.}

\item{homeOwner}{Is the person a home-owner?}

\item{sinRaMax}{Maximum fortnightly entitlement to rent-assistance for singles.}

\item{sinMaxRate}{Pension maximum base rate, per fortnight, for singles.}

\item{sinAssHomeThr}{Assets threshold for single home-owners.}

\item{sinAssNonHomeThr}{Assets threshold for singles who don't own a home.}

\item{pensAssTpr}{Taper rate for assets test.}

\item{sinIncThr}{Income threshold for income test.}

\item{sinSupTot}{Pension supplement, total amount, per fortnight for a single.}

\item{sinSupMin}{Pension supplement, minimum amount, per fortnight for a single.}
}
\description{
Value of age pension (for singles only).
}
