% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BalanceAfterRealLevelPayments}
\alias{BalanceAfterRealLevelPayments}
\title{Balance after n real level payments}
\arguments{
\item{nominal_payment}{An integer representing the payment (in nominal terms with respect to \code{balance}).}

\item{n}{Number of periods (years).}

\item{balance}{The super balance prior to retirement.}

\item{r_earnings}{The rate of earnings growth.}

\item{cpi}{The rate of CPI growth.}

\item{inArrears}{(logical) Are the payments made at the end of 12 months (\emph{i.e.} after the balance has grown)?}
}
\description{
Super balance after level payments.
}
