% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-income_transition_table.R
\docType{data}
\name{income_transition_table}
\alias{income_transition_table}
\title{Income Transition Table}
\format{Data.table of 10,000 observations:
\describe{
\item{from,to}{Lifetime income percentiles from and to.}
\item{value}{Probability of moving from \code{from} to \code{to}.}
\item{cumvalue}{Cumulative probability, by \code{from}.}
}}
\usage{
income_transition_table
}
\description{
Probability of moving from one decile to another.
}
\keyword{datasets}
