% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IncomeTax}
\alias{IncomeTax}
\title{Income tax}
\arguments{
\item{income}{The persons's income.}

\item{thresholds}{Income tax brackets}

\item{rates}{Marginal tax rates.}
}
\description{
Income tax for a single worker.
}
