% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lifetime_pachinko_density.R
\docType{data}
\name{lifetime_pachinko_density}
\alias{lifetime_pachinko_density}
\title{Lifetime 'Pachinko' Density}
\format{A \code{data.table} of 800,000 variables (number of age groups \eqn{{}\times 100}):
\describe{
\item{\code{Percentile}}{The initial percentile.}
\item{\code{id}}{ID representing a unique person.}
\item{\code{simulation}}{ID representing a unique simulation.}
\item{\code{AWOTE_average}}{Average lifetime income, expressed in AWOTE.}
}}
\usage{
lifetime_pachinko_density
}
\description{
Data table reflecting lifetime densities of incoem
}
\keyword{datasets}
