% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RealPensionLevel}
\alias{RealPensionLevel}
\title{Real level of pension payments giving starting retirement conditions}
\arguments{
\item{n}{The number of periods.}

\item{balance}{The balance at retirement.}

\item{r_earnings}{The rate of earnings growth (\code{r_earnings = 0.065} means 6.5\%).}

\item{cpi}{The rate of CPI.}

\item{inArrears}{Are the payments paid in arrears (\emph{i.e.} before the balance has grown through \code{r_earnings}.)}
}
\description{
The value of a cpi-adjusted pension in nominal terms.
}
