% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement_rate.R
\name{replacement_rate}
\alias{replacement_rate}
\title{Replacement rate}
\usage{
replacement_rate(wage_ratio, policy = "current")
}
\arguments{
\item{wage_ratio}{AWOTE multiple.}

\item{policy}{A string representing a policy world.}
}
\value{
The replacement rate.
}
\description{
Replacement rate given initial conditions.
}
