% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_balance_at_retirement.R
\name{super_balance_at_retirement}
\alias{super_balance_at_retirement}
\title{Super balance at retirement}
\usage{
super_balance_at_retirement(starting_age = 30L, retirement_age = 70L,
  start_fy = "2015-16", wage.method = "Constant \% of AWOTE",
  wage.method.control = list(AWOTE = 1), starting_balance = 0)
}
\arguments{
\item{starting_age}{An integer (thus \code{30L} not \code{30}) given the starting year of the individual.}

\item{retirement_age}{An integer. Youngest age of retirement age, not oldest age of work.}

\item{start_fy}{A financial year corresponding to \code{starting_age}.}

\item{wage.method}{One of \code{"Constant \% of AWOTE"} or \code{"Lifetime percentile"}.}

\item{wage.method.control}{A \code{list}: \describe{
\item{\code{"Constant \% of AWOTE"}}{\describe{
   \item{\code{AWOTE.start}}{the AWOTE at \code{starting_age};}
   \item{\code{AWOTE.multiple}}{the multiple of AWOTE to use.}
   }}
\item{\code{"Lifetime percentile"}}{\code{percentile} the percentile to use.}
}
Using one set of controls in the wrong \code{wage.method} is a warning.}

\item{starting_balance}{The balance of the super account when the person's age is \code{starting_age}.}
}
\description{
Determine the superannuation balance at the commencement of retirement of a single individual.
}

