% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_guarantee.R
\name{super_guarantee}
\alias{super_guarantee}
\title{The rate of the superannuation guarantee}
\usage{
super_guarantee(fy_year)
}
\arguments{
\item{fy_year}{The financial years for which the super guarantee is desired.}
}
\value{
The super guarantee for that year. See Details.
}
\description{
The rate of the superannuation guarantee
}
\details{
Currently returns 9.5\% for all years.
}

