% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SuperBalanceViaPachinko}
\alias{SuperBalanceViaPachinko}
\title{Balance via Pachinko}
\arguments{
\item{age}{Age of individual (will be rounded down to nearest 5).}

\item{percentile}{Income percentile.}

\item{starting_year}{Year of initial.}

\item{retirement_age}{Age at retirement.}

\item{AWOTE_prop}{Proportion of AWOTE.}
}
\description{
Balance of super via a machine.
}
