% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_life_remuneration.R
\name{working_life_remuneration}
\alias{working_life_remuneration}
\title{Working life remuneration}
\usage{
working_life_remuneration(starting_age = 30L, retirement_age = 70L,
  start_fy = "2015-16", wage.method = c("Constant \% of AWOTE",
  "Lifetime percentile"), wage.method.control = list(AWOTE.start = 78200,
  AWOTE.multiple = 1, percentile = NULL),
  wage.inflator = c("grattan::wage_inflator()", "Treasury"),
  forecast.series = "mean", forecast.level = 95)
}
\arguments{
\item{starting_age}{An integer (thus \code{30L} not \code{30}) given the starting year of the individual.}

\item{retirement_age}{An integer. Youngest age of retirement age, not oldest age of work.}

\item{start_fy}{The financial year corresponding to \code{starting_age}.}

\item{wage.method}{One of \code{"Constant \% of AWOTE"} or \code{"Lifetime percentile"}.}

\item{wage.method.control}{A \code{list}: \describe{
\item{\code{"Constant \% of AWOTE"}}{\describe{
   \item{\code{AWOTE.start}}{the AWOTE at \code{starting_age};}
   \item{\code{AWOTE.multiple}}{the multiple of AWOTE to use.}
   }}
\item{\code{"Lifetime percentile"}}{\code{percentile} the percentile to use.}
}
Using one set of controls in the wrong \code{wage.method} is a warning.}

\item{wage.inflator}{One of \code{"grattan::wage_inflator()"} or \code{"Treasury"}.}

\item{forecast.series}{See \code{"grattan::wage_inflator()"}.}

\item{forecast.level}{See above.}
}
\value{
A \code{data.table} of information about the individual's remuneration from \code{starting_age} till the year before \code{retirement_age}.
}
\description{
Working life remuneration
}
