% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AverageWorkingLifeIncome}
\alias{AverageWorkingLifeIncome}
\title{Working Life Income}
\arguments{
\item{prop_of_AWOTE_start_year}{The proportion of AWOTE of the subject in the \code{starting_year}.}

\item{wage_method}{'Constant prop AWOTE'.}

\item{starting_age}{The age at which the model begins. Set to 30 by default, rather than a younger age
to account for time out of the workforce.}

\item{deflator}{What deflator should be used?}

\item{Starting}{age.}

\item{Retirement}{age.}

\item{AWOTE_in_starting_year}{The salary in starting year.}

\item{short_run_wage_growth}{A vector of wage indices.}

\item{long_run_wage_growth}{Rate of wage growth beyond short_run_wage_growth.}

\item{starting_year}{The year to begin. Used for determining superannuation guarantee.}
}
