% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReplacementRate}
\alias{ReplacementRate}
\title{Replacement rate}
\arguments{
\item{starting_age}{(Default: 30) The age in the first year of the model. Integer.}

\item{retirement_age}{(Default: 70) Retirement age for the model. Integer.}

\item{wage_method}{Constant proportion of AWOTE.}

\item{AWOTE_starting_year}{The AWOTE for the starting year.}
}
\description{
Replacement rate.
}
